+++ core/domain/interface/ad/ad_types.go
@@ -19,8 +19,8 @@ type (
 
 	//  超链接
 	HyperLink struct {
-		Id      int64  `db:"id"`
-		AdId    int64  `db:"ad_id"`
+		Id      int32  `db:"id"`
+		AdId    int32  `db:"ad_id"`
 		Title   string `db:"title"`
 		LinkUrl string `db:"link_url"`
 	}
@@ -33,10 +33,10 @@ type (
 	// 广告图片
 	Image struct {
 		// 图片编号
-		Id int64 `db:"id" auto:"yes" pk:"true"`
+		Id int32 `db:"id" auto:"yes" pk:"true"`
 
 		// 广告编号
-		AdId int64 `db:"ad_id"`
+		AdId int32 `db:"ad_id"`
 
 		// 图片标题
 		Title string `db:"title"`
@@ -63,13 +63,13 @@ type (
 		GetEnabledAdValue() ValueGallery
 
 		// 保存广告图片
-		SaveImage(v *Image) (int64, error)
+		SaveImage(v *Image) (int32, error)
 
 		// 获取图片项
-		GetImage(id int64) *Image
+		GetImage(id int32) *Image
 
 		// 删除图片项
-		DelImage(id int64) error
+		DelImage(id int32) error
 	}
 )
