+++ core/domain/interface/merchant/merchant_rep.go
@@ -16,7 +16,7 @@ type IMerchantRep interface {
 	CreateMerchant(*Merchant) IMerchant
 
 	// 创建会员申请商户密钥
-	CreateSignUpToken(memberId int64) string
+	CreateSignUpToken(memberId int32) string
 
 	// 根据商户申请密钥获取会员编号
 	GetMemberFromSignUpToken(token string) int64
@@ -25,19 +25,19 @@ type IMerchantRep interface {
 	GetMerchantsId() []int64
 
 	// 获取商户
-	GetMerchant(id int64) IMerchant
+	GetMerchant(id int32) IMerchant
 
 	// 获取合作商主要的域名主机
-	GetMerchantMajorHost(mchId int64) string
+	GetMerchantMajorHost(mchId int32) string
 
 	// 保存
-	SaveMerchant(*Merchant) (int64, error)
+	SaveMerchant(*Merchant) (int32, error)
 
 	// 获取账户
-	GetAccount(mchId int64) *Account
+	GetAccount(mchId int32) *Account
 
 	// 获取销售配置
-	GetMerchantSaleConf(mchId int64) *SaleConf
+	GetMerchantSaleConf(mchId int32) *SaleConf
 
 	// 保存销售配置
 	SaveMerchantSaleConf(v *SaleConf) error
@@ -49,40 +49,40 @@ type IMerchantRep interface {
 	SaveApiInfo(d *ApiInfo) error
 
 	// 获取API信息
-	GetApiInfo(mchId int64) *ApiInfo
+	GetApiInfo(mchId int32) *ApiInfo
 
 	// 根据API编号获取商户编号
 	GetMerchantIdByApiId(apiId string) int
 
 	// 获取键值
-	GetKeyValue(mchId int64, indent string, k string) string
+	GetKeyValue(mchId int32, indent string, k string) string
 
 	// 设置键值
-	SaveKeyValue(mchId int64, indent string, k, v string, updateTime int64) error
+	SaveKeyValue(mchId int32, indent string, k, v string, updateTime int64) error
 
 	// 获取多个键值
-	GetKeyMap(mchId int64, indent string, k []string) map[string]string
+	GetKeyMap(mchId int32, indent string, k []string) map[string]string
 
 	// 检查是否包含值的键数量,keyStr为键模糊匹配
-	CheckKvContainValue(mchId int64, indent string, value string, keyStr string) int
+	CheckKvContainValue(mchId int32, indent string, value string, keyStr string) int
 
 	// 根据关键字获取字典
-	GetKeyMapByChar(mchId int64, indent string, keyword string) map[string]string
+	GetKeyMapByChar(mchId int32, indent string, keyword string) map[string]string
 
 	//获取等级
-	GetLevel(mchId, levelValue int64) *MemberLevel
+	GetLevel(mchId, levelValue int32) *MemberLevel
 
 	// 获取下一个等级
-	GetNextLevel(mchId, levelVal int64) *MemberLevel
+	GetNextLevel(mchId, levelVal int32) *MemberLevel
 
 	// 获取会员等级
-	GetMemberLevels(mchId int64) []*MemberLevel
+	GetMemberLevels(mchId int32) []*MemberLevel
 
 	// 删除会员等级
-	DeleteMemberLevel(mchId, id int64) error
+	DeleteMemberLevel(mchId, id int32) error
 
 	// 保存等级
-	SaveMemberLevel(mchId int64, v *MemberLevel) (int64, error)
+	SaveMemberLevel(mchId int32, v *MemberLevel) (int32, error)
 
 	//
 	//  //修改线下支付利润
