+++ core/domain/interface/merchant/shop/shop.go
@@ -133,7 +133,7 @@ var (
 
 type (
 	IShop interface {
-		GetDomainId() int64
+		GetDomainId() int32
 
 		// 商店类型
 		Type() int
@@ -145,7 +145,7 @@ type (
 		SetValue(*Shop) error
 
 		// 保存
-		Save() (int64, error)
+		Save() (int32, error)
 
 		// 数据
 		Data() *ShopDto
@@ -182,8 +182,8 @@ type (
 
 	// 商店
 	Shop struct {
-		Id         int64  `db:"id" pk:"yes" auto:"yes"`
-		MerchantId int64  `db:"mch_id"`
+		Id         int32  `db:"id" pk:"yes" auto:"yes"`
+		MerchantId int32  `db:"mch_id"`
 		ShopType   int    `db:"shop_type"`
 		Name       string `db:"name"`
 		State      int    `db:"state"`
@@ -193,8 +193,8 @@ type (
 
 	// 商店数据传输对象
 	ShopDto struct {
-		Id         int64
-		MerchantId int64
+		Id         int32
+		MerchantId int32
 		ShopType   int
 		Name       string
 		State      int
@@ -206,7 +206,7 @@ type (
 	// 商城
 	OnlineShop struct {
 		// 商店编号
-		ShopId int64 `db:"shop_id" pk:"yes" auto:"no"`
+		ShopId int32 `db:"shop_id" pk:"yes" auto:"no"`
 		// 通讯地址
 		Address string `db:"addr"`
 		// 联系电话
@@ -234,7 +234,7 @@ type (
 	// 门店
 	OfflineShop struct {
 		// 商店编号
-		ShopId int64 `db:"shop_id" pk:"yes" auto:"no"`
+		ShopId int32 `db:"shop_id" pk:"yes" auto:"no"`
 
 		// 联系电话
 		Tel string `db:"tel"`
