+++ core/domain/interface/merchant/shop/shop_rep.go
@@ -10,31 +10,31 @@ package shop
 
 type (
 	IShopRep interface {
-		SaveShop(*Shop) (int64, error)
+		SaveShop(*Shop) (int32, error)
 
 		// 商店别名是否存在
-		ShopAliasExists(alias string, shopId int64) bool
+		ShopAliasExists(alias string, shopId int32) bool
 
 		// 获取商店值
-		GetValueShop(mchId, shopId int64) *Shop
+		GetValueShop(mchId, shopId int32) *Shop
 
 		// 获取商户所有商店
-		GetShopsOfMerchant(mchId int64) []Shop
+		GetShopsOfMerchant(mchId int32) []Shop
 
 		// 删除线上商店
-		DeleteOnlineShop(mchId, shopId int64) error
+		DeleteOnlineShop(mchId, shopId int32) error
 
 		// 删除线下门店
-		DeleteOfflineShop(mchId, shopId int64) error
+		DeleteOfflineShop(mchId, shopId int32) error
 
 		// 获取线上商店
-		GetOnlineShop(shopId int64) *OnlineShop
+		GetOnlineShop(shopId int32) *OnlineShop
 
 		// 保存线上商店
 		SaveOnlineShop(v *OnlineShop, create bool) error
 
 		// 获取线下商店
-		GetOfflineShop(shopId int64) *OfflineShop
+		GetOfflineShop(shopId int32) *OfflineShop
 
 		// 保存线下商店
 		SaveOfflineShop(v *OfflineShop, create bool) error
